Popup to confirm if you want to delete a question or not

+++ web/src/pages/OwnerQuestionsPage.js
@@ -1,4 +1,5 @@
 import React, { useEffect } from 'react'
+import Swal from 'sweetalert2'
 import { connect } from 'react-redux'
 
 import { fetchOwnerQuestions, deleteQuestion } from '../actions/questionActions'
@@ -16,7 +17,26 @@ const OwnerQuestionsPage = ({ dispatch, loading, questions, hasErrors, redirect,
     }, [redirect, dispatch, userId]);
 
     const onDelete = (id) => {
-        dispatch(deleteQuestion(id))
+
+        Swal.fire({
+            title: 'Estas seguro de eliminar la pregunta?',
+            text: "Se eliminara permanentemente",
+            icon: 'Advertencia',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Si, eliminar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                dispatch(deleteQuestion(id))
+                Swal.fire(
+                    'Eliminada',
+                    'Tu pregunta ha sido eliminada.',
+                    'Eliminacion exitosa'
+                )
+            }
+        })
+
     }
